+++ mod/racing_plus/Sprites.lua
@@ -191,23 +191,27 @@ end
 function Sprites:GetScreenCenterPosition()
   -- Local variables
   local shape = g.r:GetRoomShape()
-  local centerOffset = (g.r:GetCenterPos()) - g.r:GetTopLeftPos()
-  local pos = g.r:GetCenterPos()
+  local centerPos = g.r:GetCenterPos()
+  local centerOffset = centerPos - g.r:GetTopLeftPos()
 
   if centerOffset.X > 260 then
-      pos.X = pos.X - 260
+    centerPos.X = centerPos.X - 260
   end
-  if shape == RoomShape.ROOMSHAPE_LBL or shape == RoomShape.ROOMSHAPE_LTL then
-      pos.X = pos.X - 260
+  if shape == RoomShape.ROOMSHAPE_LTL or -- 9
+     shape == RoomShape.ROOMSHAPE_LBL then -- 11
+
+    centerPos.X = centerPos.X - 260
   end
   if centerOffset.Y > 140 then
-      pos.Y = pos.Y - 140
+    centerPos.Y = centerPos.Y - 140
   end
-  if shape == RoomShape.ROOMSHAPE_LTR or shape == RoomShape.ROOMSHAPE_LTL then
-      pos.Y = pos.Y - 140
+  if shape == RoomShape.ROOMSHAPE_LTL or -- 9
+     shape == RoomShape.ROOMSHAPE_LTR then -- 10
+
+    centerPos.Y = centerPos.Y - 140
   end
 
-  return Isaac.WorldToRenderPosition(pos, false)
+  return Isaac.WorldToRenderPosition(centerPos, false)
 end
 
 -- This clears the graphics that should only appear in the starting room