+++ mod/racing_plus/UseCard.lua
@@ -6,7 +6,9 @@ local PostItemPickup = require("racing_plus/postitempickup")
 
 function UseCard:Main(card)
   -- Display the streak text (because Racing+ removes the vanilla streak text)
-  if card ~= Card.RUNE_BLANK then -- 40
+  if card == Card.RUNE_BLANK then -- 40
+    g.run.streakForce = true
+  else
     -- We ignore blank runes because we want to show the streak text of the actual random effect
     g.run.streakText = g.itemConfig:GetCard(card).Name
     g.run.streakFrame = Isaac.GetFrameCount()